MAINT: reorganize shared linear assignment tests
As noted by @rgommers in https://github.com/scipy/scipy/pull/12541,
`scipy.optimize` already depends on `scipy.sparse.csgraph` through
`scipy.sparse`, so we can add the shared cases to `scipy.sparse.csgraph`
without introducing any additional cross-package dependencies, just like
we can add the test case using both solvers to `scipy.optimize` without
adding dependencies.

So that's what we do here. Note that this introduces some very slight
duplication between the two test modules, as both tests will have to
eventually make use of the newly introduced
`linear_sum_assignment_assertions` for the same inputs. It would be
possible to avoid this by combining the test cases and the assertions
directly, but that would come at the cost of not being able to make
simple use of `pytest.mark.parametrize`.
